// public/apps/the_peoples_edh_bracket/the-peoples-edh-bracket-client/src/types/bindings/Deck.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CardWithGlobalPoints } from "./CardWithGlobalPoints";
import type { DeckMaindeckEntry } from "./DeckMaindeckEntry";

export type Deck = { commanders: Array<CardWithGlobalPoints>, maindeck: Array<DeckMaindeckEntry>, };